Fix package building for Debian + PE
This commit fixes a few issues that were causing issues when attempting
to build puppetdb for Puppet Enterprise.

The Rakefile now prepends /opt/puppet/bin as your first path search to
ensure the tools being used are from Puppet Enterprise.

The templates and erb files had some minor errors and ommissions that
needed to be fixed to build out packages for pe-puppetdb and
pe-puppetdb-terminus on Debian/Ubuntu.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>