(#14613) Don't fail when collecting exec resources
We were adding aliases for exec resources based on their command, which
is the namevar. This is so that we can properly lookup resources
referred to by relationship metaparameters when producing relationship
edges. However, exec resources aren't isomorphic, meaning they aren't
unique based on their namevar. This means we don't need to add those
aliases, because it's illegal to specify a relationship to an exec
resource using its command. It also means we were adding non-unique
aliases, which were evaluated on the agent after collecting the
resources, and could cause conflicts.