Tolerate leading whitespace in puppetdb.conf
The INI file parser included in Puppet treats lines beginning with
whitespace to be continuation lines, adding them to the value of the
previous setting. This causes indentation to break puppetdb.conf, which
is undesired. For example:

[main]
  server = puppetdb.example.com
  port = 8081

This will cause the value of the server setting to be
"puppetdb.example.com\n  port = 8081", and the port setting will not be
set by the file, and will just default to 8080. This is definitely not
our desired behavior, and indentation is something we want. So the
INI parser has been replaced with a simple custom parser which handles
this the way we want it.