Fix init script for EL based PE installations
Previously, we were grepping on username in the find_my_pid function of
the init script.  This was fine if the username was <= 8 characters, but
since pe-puppetdb is not, 'ps -ef' would fall back to using UID in the
processes table, thus causing the function to fail.  This in turn led to
empty pid files and puppetdb potentially starting lots of times.

We now grep the jarfile name rather than user to confine the process
listing.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>