Support a "version" file instead of `git describe`
When running from source but not a git repository (for instance, after
copying the code to another machine to build a package..), lein fails
because it can't find the version. This allows the use of a file called
"version" which contains the version. If this file exists, it will be
used preferentially over `git describe`.