(#14873) Fix ruby 1.8 utf-8 encoding issue that was causing checksum failures
Certain byte sequences that are not considered valid in UTF-8 were
not being detected by our string sanitization methods.  This was
causing checksum errors between ruby and clojure.

This commit adds a spec test and an acceptance test that illustrate
a particular sequence of bytes that was known to cause this problem,
and changes our string-cleaning code so that it does a better job
of detecting invalid byte sequences.