Upgrade to Ring 1.1.1
Ring 1.1.1 contains the necessary changes to its servlet adapter to allow for
inspection of client-supplied SSL certificates. Upgrading also requires an
upgrade to Jetty 7. This required some changes in the handful of monkeypatches
we made to ring-jetty-adapter, though the new versions of ring obviate some of
them (ring now includes support for client-auth, so we no longer need to roll
our own).

The new version of Jetty pulls in its own version of various logging libraries,
and one of these conflicts with the version required by some of the other
libraries we're using. To resolve this conflict, I've excluded the problematic
library from each dependency that tries to fold it in itself, and instead
explicitly require a specific version.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>