(#14947) Restrict acceptable client certificates by CN
This patchset implements an optional whitelist for HTTPS clients. If
enabled by the user, we validate that the CN of the supplied client
certificate exactly matches an entry in the whitelist. This feature is
off by default.

This is implemented as ring middleware and a single, additional
configuration option. For plain HTTP connections, all clients are
allowed (as no client certificate will be present).

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>