(#14837) Auto-insert new keystore/truststore pw into jetty.ini
Previously, the ssl-setup script would generate new keystore and
truststore files but would fail to update an existing jetty.ini file
with the updated password to those files. Thus, re-running the setup
script would leave a user's puppetdb installation in a broken state.

This patch checks to see if there's a jetty.ini file, and that it
contains the settings for keystore and truststore passwords. If so,
then we update the passwords contained within. If not, we output a
message telling the user to manually update their password settings,
as we can't do it automatically.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>