Improve test setup / teardown stuff
This corresponds to pull req #247 in the acceptance harness.
That pull req gives us improved hooks for setup / teardown of
test state, and this pull req takes advantage of those hooks.

With this commit it becomes possible to run the puppetdb
acceptance tests several times in series without having to
reset your VMs between each run.