use "node_name" when activating / deactivating node
This test was relying on the assumption that the
"to_s" method of the "Host" class in the acceptance
framework would return the exact same value as the
agent's "node_name_value".  This isn't always the
case, so this commit changes the test to use
the node_name value explicitly.

(Relies on pull req #248 for the acceptance framework.)