Condition remove_previous_config based on vmrun
If you are running the acceptance tests with
a --vmrun option, you don't need the step
that removes old configuration files during
startup (and in fact that step won't work
if your VM is not running and accepting
ssh connections).  This commit adds a conditional
so that we don't try to do that if you are
using --vmrun.