Fix broken test
There was a recent change in the acceptance framework that
caused "result.host" ("result" being the return value from an exec
command) from an instance of Host to just a String containing the
hostname.  We had one test that was relying on the old behavior;
this changes it to be compatible with the new behavior.