Safely ordered and more thorough catch-all logging
We now log to stdout first, prior to using the system logger. This
avoids a situation where the system logger is misconfigured (or is
otherwise broken) and attempts to log to it would themselves trigger
an exception.

This also allows us to ensure that the information logged to stdout is
properly flushed prior to forwarding the exception to the system
logger. That means we can reduce the likelihood of interleaved output
on stdout.

Finally, we replace the call to `printStackTrace` with the
`print-cause-trace` function that's part of the clojure.stacktrace
library.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>