Simplify function usage
There are cases where threading is used, but doesn't necessarily add
anything by doing so. In the cases where a single function is called on
the data, it reads cleaner to just call that function on the data you
hold.

Also update use of greater-than function as the check is simply for
positive values.