Verify config path before starting up
Prior to this commit, if the config path that the user
specified did not exist or was not readable, we'd silently
ignore it and then fail a bit later with a confusing
error message about not finding a "vardir" setting.

This commit adds a bit of explicit checking to make
sure that the specified config path exists and is
readable before we try to parse the ini file(s).