Work around non-string resource titles
It's possible in some cases for Puppet to generate a resource whose
title isn't a string. However, since the generated edges refer to the
resource using a string title, we end up with a mismatch. Now we will
stringify all resource titles on the way out. In future, Puppet should
do this for us.