(#14688) Improve stdout/stderr handling for redhat init script
Prior to this commit, the redhat init script was keeping stdout/stderr
opening when you called "service puppetdb stop".  This resulted
in some undesirable behavior; starting the service over an ssh
connection would not release the ssh connection, errors would
appear on the console rather than in the log file, etc.

This commit simply changes the init script to call "daemon" and
to redirect stdout/stderr to a file.  In local testing, this
appears to solve both of the reported problems (ssh hangs and console
errors).