Stop recursive symlinks from `rake install`
We were using `ln -sf`, which will symlink a directory into itself when
the source is a directory and the destination is already a symlink to
the source. This is what happens when `rake install` is run twice. Now
we use `ln -sfT`, which treats the destination as a regular file, and
won't traverse it when considering where/whether to make the symlink.