(#15281) Add ability to toggle between database backends
This commit allows the acceptance tests to be configured
to use either the embedded database OR postgres.  It
introduces some "options files" (which rely on pull req
number 259 on the acceptance framework)--these are used
to provide default values for various options that
we would normally specify from the command line; they
are also used to introduce the "puppetdb_database" option,
which is what we will use to toggle between the different
database backends.