Add SRPMS directory creation to rake:SRPM task
Some configurations of rpmbuild won't make the SRPMS directory automatically
when building an RPM, so this commit adds a mkdir_p to the rake task to make
the directory before running rpmbuild. Otherwise, the rpmbuild call fails when
rpmbuild can't write the src rpm to the SRPMS directory (because it doesn't
exist).