Add support for a --debug flag, and a startup script
This commit adds support for a new command-line flag:
--debug.  For now, this flag only affects logging;
it forces a console logger and ensures that the
log level is set to DEBUG.  The option is also
added to the main config hash so that it can potentially
be used for other purposes in the future.

This commit also adds a shell script, "puppetdb-services",
which can be used to launch the services from the
command line.  This script will be packaged (in /usr/sbin)
along with the puppetdb-ssl-setup script, and will
hopefully be useful in helping users troubleshoot
problems on their systems.