Configure yum and apt to use a proxy during acceptance
This commit adds a setup step to our acceptance testing
which configures yum or apt (depending on the puppetdb
host's OS) to use a proxy server.  Hopefully this
will shorten the run time of our acceptance runs and
make us less susceptible to problems related to
packet loss between the office and the outside world
when our network gets shaky.