Fix deb packaging so that service is restarted on upgrade
Prior to this commit, when you ran a debian package upgrade,
the puppetdb service would be stopped but would not be
restarted.  This commit removes some DEBHELPER stuff in
order to give us some more explicit control over the
lifecycle.  It adds a "prerm" script to shut down the
service, and adds a block to the "postinst" script to make
it explicitly start the service again.