Remove use of current_timetamp
For time-based interactions with the database, we need to standardize on
one of two approaches:

1) Always use the database's notion of time
2) Always use the JVM's notion of time

Mixing these two things can result in some seriously strange race
conditions between the daemon's idea of time (like when we stamp
messages with a received-time annotation) and the database's idea of
time.

This patchset removes any lingering use of `current_timetamp`, replacing
such usage with JVM time.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>