(#15696) Add setting for restricting number of results for a resource query
This commit introduces a setting called 'resource-query-limit',
which will cause resource queries to fail if they would return
more results than the specified value.

It also adds some utility functions to the jdbc namespace to
allow this sort of 'limit' behavior to be used elsewhere if
we decide to do so.