Add/improve support for logging SQL queries
Prior this this commit, our support for logging raw SQL statements
was not well-defined (and perhaps not working).  This commit
introduces and documents two settings that you can include in
your database.ini file to control logging of SQL.

* Setting 'log-statements' to true (which is now the default) means
  that all SQL statements will be logged at DEBUG.  Because we
  leverage log4j's dynamic logging properties, this means that you
  can change the log4j.properties on a running service and set the
  log level to DEBUG if you wish to start logging queries without
  restarting the service.

* 'log-slow-queries' defines the maximum number of seconds that
  a query can execute before it is considered "slow".  "slow" queries
  are logged at the WARN log level.

This commit also introduces a new "internal" namespace for our jdbc
functions... this is intended to provide a distinction between
functions that are pseudo-public and functions that are pseudo-private
and should only be used internally.