Fix issues with using new version of jetty.ini on upgrade
Previously the user could choose to use the "new" version of jetty.ini,
which was a template including a placeholder password. During a regular
install, we run puppetdb-ssl-setup, which replaces the placeholder with
the actual password. However, during an upgrade, we weren't running it.
Now, we will always run the script, and determine whether or not to
generate a keystore/truststore/password based on the existence of the
password file. If it already exists, we will update the jetty.ini to use
it.