Add packaging support for Fedora 17/ruby 1.9
This commit adds support for packaging for ruby
1.9. It modifies the @plibdir path based on the
ruby version. RUBY_VER can be passed in as an
environment variable, and if none is passed,
RUBY_VER defaults to the ruby on the local host
as reported by facter. As is currently the case,
we use the sitelibdir in ruby 1.8, and with this
commit use vendorlibdir for 1.9. Fedora 17 ships
with 1.9, so we use this to test for 1.9 in the
spec file. Fedora 17 also ships with open-jdk 1.7,
so this commit updates the Requires to 1.7 for
fedora 17.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>