(maint) Update puppetdb packaging to correctly handle RCs
Previously the puppetdb packaging would create packages with versions like
1.0.0rc1, which is a problem for rpm and deb packages because 1.0.0 < 1.0.0rc1,
so 1.0.0 would never be installed after 1.0.0rc1 had been. This commit updates
the packaging tasks to correctly build packages named to our RC standard, such
as 1.0.0-0.1rc1, which will be less than 1.0.0-1, the final version. The commit
adds instance variables for @debversion, @rpmversion, @origversion,
@debrelease, and @rpmrelease as well as helper methods to populate those
variables. It also updates the changelog template for debian and the spec file
template for redhat to correctly use the new variables.  This also includes
updates to the acceptance tests regarding version checks to correctly check the
full version and release of rpms and debs.