Add configurable update server and rework error handling
The `update-server` setting can now be set in the config file in the
[global] section in order to change the location to use for checking for
updates. This also allows the user to entirely disable this behavior by
using a non-functional route, if they absolutely cannot tolerate the
request.

When someone makes a request to /version/latest, we no longer swallow
exceptions, as this was a deliberate request. Additionally, we now *do*
swallow exceptions when checking for updates in services, though we log
them at debug level for the sake of not throwing away information the
user may not care about.