(#15903) Detect invalid multi-byte sequences from 0xC0, 0xC1, 0xF5-0xFF
Prior to this commit, our UTF-8 sanitizer was not recognizing
multi-byte sequences starting with 0xC0, 0xC1, or anything in
the range of 0xF5-0xFF invalid.  Java/Clojure *do* consider
these invalid (which matches up with the explanation on the
UTF-8 page on wikipedia).  This commit adds detection for
those byte sequences.