Duplicate collected resources should cause a failure
Prior to this commit, the puppetdb resource terminus
was not setting a value for "collector_id" on collected
resources.  This field is used by puppet to detect
duplicate resources (exported by multiple nodes) and
cause a run to fail.

This commit adds code to set the collector id based on
node name + resource title + resource type, and adds
tests to verify that a puppet run will fail if it
collects duplicate instances of the same resource from
different exporters.