Add support for a test configuration file
This commit adds support for a configuration file to be used
during unit test runs.  At present, the only thing that you
can define in the config file is a function that returns
a map of database settings.  There is a default configuration
file provided, which simply returns the same database settings
that we were using before (an in-memory hsql database with
a random UUID).  However, it also provides instruction for
creating a local override of the default config file, including
an example of how to configure your tests to run against
postgres instead of hsql.

This commit also tweaks the with-test-db fixture to call
a database-cleanup function prior to running the database
migration.  This cleanup function drops all of the tables
if they exist, so that each test can start with a fresh db.