(#15645) Daemon should terminate upon OOM error
Though we can potentially catch an OOM error and do something about it,
it’s not guaranteed that the error will occur in a thread in which we
can interpose our own logic. But certainly we shouldn’t keep the process
going if an OOM renders it effectively inoperative...better to just kill
the process altogether than to leave a zombie around.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>