Correct relationships between several tables
The edges table now uses the metadata hash to refer to resources, rather
than the params hash. The resource_metadata table no longer refers to
resource_params, which was silly because metadata is deduped and not
associated with particular params. The reference to resource_params has
been correctly moved to catalog_resources.