Using the string/join function for non-lazy evaluation
The use of interpose is useful if we're expecting to produce a lazy seq
seperated by our delimeter. In this case, we were taking this lazy
sequence and converting it into a string (non lazy) anyway.

Rather than add this implicit complexity, use clojure.string/join to do
the interposing of the delimeter using a StringBuilder recursing over
the sequence we provide it.