Use a fixed revision of puppetdb module for acceptance tests
Prior to this commit, the acceptance tests were using the HEAD
revision of my puppetdb module to do setup during acceptance tests.
This was fragile and broke when I merged in some changes to the
module.

This commit simply changes the acceptance setup code to use a fixed
revision of the puppetdb module instead of HEAD.

As soon as we've released the module to the forge, we need to switch
over and use that instead of checking it out from git.