DRY up some pre/post-conditions w/rt SQL query format
We have a somewhat bulky chunk of predicates that we
use to validate that SQL queries are being passed around
to our functions in the expected type of data structure.
Prior to this commit, we had copied and pasted those
predicates in several places; this introduces a
query/utils namespace and refactors the existing usages
to use the new single function.