Add a v2 facts API with support for getting fact values
This implements a generic /v2/facts?query=[...] endpoint, which can be
used to, among other things, get all values of a given fact on all nodes
(or set of facts on a subset of nodes). The format returned is:

    [{:node <node> :fact <fact name> :value <fact value>}
     ...
     {:node <node> :fact <fact name> :value <fact value>}]

/v2/facts/<node> can still be used to get the facts for a given node (as
can /v2/facts?query=["=", ["node", "name"], <node>]), though the results
will now be returned in the above format.