Disable redhat firewall management during acceptance
This commit changes the acceptance setup such that it
does *not* attempt to manage redhat firewalls during
tests.  This seems necessary because our current
centos5 jenkins VMs have a weird configuration where
iptables is installed but not running, and so the
firewall module fails when it tries to manage the
firewall.

Hopefully this commit will get the tests passing for now,
but the correct longer-term solution for this is to
make the firewall module more resilient and/or get our
jenkins nodes into a more normal state so that we can
re-enable this, because it would be useful for it to be
included in the tests.