Add 'purge_after_run' setting for acceptance
This new setting allows us to flag whether or
not to remove packages and do other exhaustive
cleanup after a run.  The previous behavior
was to always do this, which was useful when
doing repeated runs of the acceptance tests on
a VM without restoring a snapshot of the VM
between runs.  This commit changes the default
behavior so that the packages are *not* removed
after a test run, which is more useful for
debugging the runs on CI; the original behavior
can still be achieved by defining the setting
via an environment variable, etc.