Acceptance: after installing, sleep until started
Some of the recent acceptance failures look like
they may just be from race conditions where
the puppetdb server wasn't fully started up
by the time the master's manifest was applied
to try to connect to it.  This commit simply
causes the install_puppetdb method to, after
installing the package, sleep until the
service is up and running.