Make retrieval of table names dynamic
Prior to this commit we had a hard-coded list of table names
that was used during testing to clear out the database.  In
order to reduce code maintenance efforts in the future, this
commit changes to a dynamic means of discovering the table
names so that the hard-coded list doesn't need to be
maintained.