Fix sourceinstall to use strings for ENV
Prior to this commit, the rake 'sourceinstall' task
was using a numeric literal as the value for an
assignment into the ENV hash, which is a special
hash in ruby that can only accept strings for its
values.  This meant that the rake sourceinstall task
was unusable.  This commit changes things over to
use strings rather than numbers.

However, this raises the question as to whether anyone
is using this for anything and whether or not we should
just delete it.  Need to talk to release about that.