(#16407) Handle trailing slashes when creating edges for file resources
This commit adds a few more tests for specifying relationships
to File resources with a different number of trailing slashes
than the resource was originally declared with.

It also adds a special-case to the logic for synthesizing
edges to file resources, to munge off the trailing slashes and
look up the file resources in the catalog via their canonical
title.