Properly find edges that point at an exec by an alias
During namevar aliasing, we end up changing the :alias parameter to
'alias' and using that for the duration (to distinguish "our" aliases
form the "original" aliases). However, in the case of exec, we were
bailing out early because execs aren't isomorphic, and not adding
'alias'. Now we will always change :alias to 'alias', and just won't add
the namevar alias for execs.