Improve edge synthesis error messages
For reverse relationships (subscribe and require), the resources in edge
errors were being printed in the wrong order. This was really confusing
when trying to figure out what went wrong. This change introduces a new
format for the error, from:

Can't synthesize edge: Notify[foo] -notifies- Notify[bar] (param notify)

to:

Invalid relationship: Notify[foo] { notify => Notify[bar] }

This uses the natural Puppet language representation of the
resource/relationship to make it very clear what's going on. Also,
"edge synthesis" isn't really a term the user needs to be familiar with,
so it's replaced with something a little more obvious.