Initial scripts / rake tasks for building test packages
This commit does the following:

1) Copy all of the hacky shell code that we are using in Jenkins
   for building our PuppetDB packages into shell scripts in our
   repo.
2) Add a rake task ('test:package') that calls them.

This is not an ideal solution, but it is a step forward.  For now,
it gives us the following benefits:

1) Allows us to track the nastiest parts of our current slew of
   Jenkins jobs in VCS, and get rid of it inside of the Jenkins
   console.
2) Allows us to build a Jenkins job that can run our two packaging
   tasks (rpm, deb) in parallel.

There is more work that should eventually be done here, including
reconciling with Moses' standardized packaging tasks, porting some
or all of this shell stuff over to Ruby, and refactoring to
DRY up some of the repetitive bits.  However, I think this is
a big enough step forward to warrant committing it in its current
state.