(#16904) Add a default body to HTTP responses when none is supplied
For every status except 405 Method Not Allowed, this is currently the
description used in RFC 2616 section 10. For 405, we provide a message
that specifies the method being used and URL requested, and explains the
problem. If we want more specific messages for other codes in the
future, it's easy to add them.