Explicitly set which version of packages to install for acceptance
This will fix an issue where, for instance, the most recent released
version is 1.0.1, but that tag only exists in 1.0.x, so when testing
against master, the development package is never installed because it
isn't "newer".

This also changes the version format specified for
PUPPETDB_EXPECTED_VERSION to exactly the version returned by `rpm -q`
and `dpkg-query`, rather than munging it. That change is because we
already need the unmunged version string to specify which version to
install.