Explicitly set which version of packages to install for acceptance
(Cherry pick of b0beb79ea17950a33438b5110c1f63fdb17b7ea0 from master)

This will fix an issue where, for instance, the most recent released
version is 1.0.1, but that tag only exists in 1.0.x, so when testing
against master, the development package is never installed because it
isn't "newer".

This also changes the version format specified for
PUPPETDB_EXPECTED_VERSION to exactly the version returned by `rpm -q`
and `dpkg-query`, rather than munging it. That change is because we
already need the unmunged version string to specify which version to
install.

Conflicts:

	acceptance/helper.rb