Reduce DB connection pool partition count
We defaulted to creating 5 partitions in our connection pooling code.
Partitions are useful only inasmuch as they reduce lock contention when
attempting to acquire a new connection. Even in high-throughput
environments, we are never bottlenecked on connection acquisition.

By defaulting to 1 partition, we reduce the number of bookkeeping
threads required. Each partition requires 5 bookkeeping threads, so by
reducing the partition count to 1 we eliminate 20 threads, and save
about 20M of resident memory on users systems.

Signed-off-by: Deepak Giridharagopal <deepak@puppetlabs.com>